Refactor plugins for OpenSesame 4.0 api
