+++ opensesame_plugins/pygaze/__init__.py
@@ -1,0 +1,2 @@
+# The name of the packages to check for updates on conda and pip
+packages = ['pygaze']
