+++ opensesame_plugins/pygaze/pygaze_stop_recording/__init__.py
@@ -1,0 +1,10 @@
+category = 'PyGaze'
+priority = 10
+help = 'manual/eyetracking/pygaze'
+icon = 'os-pygaze_stop_recording'
+controls = [
+  {'type': 'line_edit',
+    'var': 'status_msg',
+    'label': 'Status message',
+    'tooltip': 'A text to use as status message'
+   }]
