--- opensesame_plugins/pygaze_start_recording/info.yaml
@@ -1,9 +1,0 @@
-category: PyGaze
-priority: 10
-help: manual/eyetracking/pygaze
-icon: os-pygaze_start_recording
-controls:
-  - type: line_edit
-    var: status_msg
-    label: Status message
-    tooltip: A text to use as status message
